starting spec language

+++ core/modules/probe.c
@@ -26,8 +26,8 @@ struct probe_priv {
 };
 
 struct report {
-    uint64_t src_addr;
-    uint64_t dst_addr;
+    uint32_t src_addr;
+    uint32_t dst_addr;
     uint64_t src_port;
     uint64_t dst_port;
     uint64_t  protocol;
